Port the complete mode-two defensive actor parent
Complete $86:F6CD-$F793 with exact timer/context routing, preserved caches
and +$64, full signed controller/role words, and native approach,
acceleration, pose, jump and final-direction ordering. Bind the B37C scratch
publication and retain the production caller path.

Verify 43 repeated native calls x 197 fields, all 81 parent starts, real
nba_tipoff_update integration, 107 legacy parent cases, and fixture mutations.
The configured suite passes across frozen-input segments: initial pre-CPU
gates and CPU prefix, exact corrected attachment block, and unchanged CPU
tail. The 63,800-row trace verifies 18,438 valid attachments, 213 strictly
classified invalid-cache rows, 2,714 exact pass frames and 138 unlocks.

Replace the old frame-774 attachment coverage pin with a pack-derived
semantic witness. Reproduce old image hashes with the preserved baseline
and inspect corrected C-only Tipoff, Mode-1 and CPU anchors; retain native
pixel rules and gameplay assertions. Record evidence and 50 newly verified
starts. Human action dispatcher $84:E2AC-$E3E9 is the next bounded scope.

No assets, packs, ROMs, images, or raw captures are included.

+++ include/nba_gameplay_ai.h
@@ -172,7 +172,7 @@ typedef struct {
     int16_t paired_velocity_x, paired_velocity_y;
     uint8_t paired_anchor_direction_raw_88;
     uint16_t paired_anchor_distance_raw_8c;
-    uint8_t paired_position_raw_92;
+    uint16_t paired_position_raw_92;
     int16_t context_anchor_x;
     uint16_t context_mode_raw_30;
     uint16_t context_flags_raw_32;
